+++ src/Core/Data/Command.cs
@@ -1,4 +1,6 @@
 ﻿
+using System.Collections.Generic;
+
 namespace AstralKeks.Workbench.Core.Data
 {
     public class Command
@@ -7,7 +9,8 @@ namespace AstralKeks.Workbench.Core.Data
         public const string Workspace = "Workspace";
 
         public string Name { get; set; }
-        public string Arguments { get; set; }
+        public List<string> Arguments { get; set; }
+
         public bool UseShellExecute { get; set; }
         public bool WaitForExit { get; set; }
         public bool NoWindow { get; set; }